Improve logging (#27)
Switched to using new dans-bag-lib instead of old version with Library of Congress domain in it.
Added a console appender so that log messages also end up in journald. This appender does not need to log the time, as journald already takes care of that.
Made sure that the bag lib only logs warnings and higher, as it contains a lot of unwarranted INFO statements.
Added SyslogIdentifier to the service unit, so that log statements of different instances of the service are easier to tell apart in journald.
Made sure that logging statements start with [<depositId>], as per the new convention introduced in dd-dataverse-ingest.